Re-aarange LocationBar component to make it re-usable
