+++ src/App.tsx
@@ -6,6 +6,7 @@ import OptimizePage from './pages/OptimizePage/OptimizePage';
 import Header from './components/Header/Header';                                                                                                                                                                                                      
 import BookingPage from './pages/BookingPage/BookingPage';
 import LandingPage from './pages/LandingPage/LandingPage';
+import LoginPage from './pages/LoginPage/LoginPage';
 
 
 const App = () => {
@@ -20,6 +21,7 @@ const App = () => {
         <Route path="/" element={ <LandingPage />} />
         <Route path="/Booking" element={ <BookingPage />} />
         <Route path="/Optimize" element={ <OptimizePage />} />
+        <Route path="/Login" element={ <LoginPage />} />
       </Routes>
     </div>
   );
